[Defect #10044] - Can no longer import from a mapped network drive to which the service user account does not have access
Upon review, decided to move the temp directory permanently to
Platform.ApplicationDataDirectory\TemporaryFiles. This will be used in all
cases. Since we can access this directory for the database, it also should
work as a temp directory.

Also made a change to publishing so that we use the directory
Platform.ApplicationDataDirectory\Publishing to store temporary files.