[Defect #10080] - Re-index starts while study import is in progress
Found bug in [c33478a7f7c1285f608d7bcce0929302182ecb2c/xian]  The logic
implemented in the changeset was invalid, and it was causing imports to
happen when reindexes were in progress.

The offending changeset was trying to introduce an issue where an import
scheduled before an In Progress Reindex was reset to pending, we wouldn't
start the import.  It ended up causing the import to start if there
was an In Progress workitem scheduled ahead of it!