[Defect #10147] - Unable to shutdown service with "stuck" retrieves [Defect #10077] - Retrieve Work Items should not be cancelable
Fix [93bbbfe23e0c6a491cfd97f453aa44b17927614e/xian] introduced this problem.
Previously, the ProgressEvent was set, and the move SCU would return, however,
a background thread was left around for the association waiting for it to timeout.
The change made it so that it would wait for the timeout before the Move SCU
would return.  With this change, we are guarenteed to send the c-cancel, and
we will wait 10 seconds for a response and the association to shut down.

Also made a general change so that a timeout occurs after 15 minutes on the
association instead of waiting 60 minutes.