[Requirement #9728] - Perform Design Reviews [Defect #10154] - Imports scheduled after a waiting re-index will run ahead of the re-index
Moved the CanStart() method in the WorkItemProcessors into a new WorkItemQuery
class.  Also moved in the support routines that were in the BaseItemProcessor.

Made WorkItemQuery create a connection, and conduct all the CanStart() checks
against the same DB connection as we're querying against.  Also, pulled in
the postponement of workitems that can't start, so they're delayed properly.
Also introduced publishing of the updates as they go along.

Also, required a small change in creating a progress structure for DICOM sends,
so that the reason the send is delaying was able to be filled in.

Still needs some updates to clean up the code a bit, but wanted to get this
much committed tonight right away.  Probably should be moving the handling
of getting entries to "delete" into this code, also...