[Defect #10151] - Imported study files are lost after re-index
Fixed small issue when scaning files, we were comparing the file name to
the Media Storage SOP Instance UID tag instead of the SOP Instance UID tag.
The issue was that these differed in the sample DICOM file, and it was causing
the files not to be reindexed.  Note tha this could only happen with badly
encoded imported files.  Files received over the network would always be
correct because the Media Storage SOP Instance UID tag is populated correctly
by us.