[Requirement #9728] - Perform Design Reviews
Added several extension methods to IQueryable<WorkItem> to query on status
values to make the code easier to understand what is happening.  The only
logical change in the code is that "Canceling" status is now considered
"active".  This was a mistake in the previous code, other work items should
wait to start running until a workitem is "canceled".