[Defect #10051] - Studies deleted during send don't disappear from the Study Browser right away
When we added in the ability to directly delete a Pending WorkItem, missed
the change to make sure this bit is reset for DeleteStudy WorkItems.

The code that processes these requests, and then runs special queries is
ugly, the code review comments point towards having to rework WorkItemService and
have an extensible way to perform "extra" processing when updates are made
to Workitems, that is WorkItem specific.  Decided to go with the simple solution
to this specific problem for now, but with the ideal that we'd refactor the
code in the future to make it work better and be more extensible.