[Requirement #9728] - Perform Design Reviews [Defect #10154] - Imports scheduled after a waiting re-index will run ahead of the re-index
The last checkin really wasn't it.  Discovered the scheduling was a bit messed
up in that I didn't allow DICOM retrieves + DICOM receives to be processed at
the same time, so the code was waiting for the retrieve workitem to complete
before the processing of the received data would start.