[Defect #10154] - Imports scheduled after a waiting re-index will run ahead of the re-index
Finally fixed the actual defect.  The CanStartNonExclusive() method was not
checking for Exclusive entries scheduled in front of it to block on.  It
now does this.  Updated the regression tests to also take this into account.
Note that this change caused some other changes to the regression tests that
were failing.