[Requirement #10229] - The workstation will provide a new Scheduling folder system, with a calendar view for scheduling patient appointments
Preliminary changes to the client code so that we can allow a folder system to have a "content view" that is different from the default items/preview view.  Basically, the idea is that a folder system can declare that it uses a custom content view, and when that folder system is active the homepage will show the custom view in place of the standard items/preview pair.

 * did some clean-up of the FolderExplorerComponent: extracted an interface IFolderExplorerComponent, so as to cleanly separate the "public API" (which is now exposed via the interface) from the "presentation model" (still exposed as public methods on the class).

 * introduced a StackedComponentContainer, which is sort of like the TabComponentContainer but without any tabs.  It's purpose is simply to allow multiple app components to be displayed within the same content area, programmatically switching between them.  This was necessary to make possible the "custom content view" idea.