[Defect #10266] - Importing single image study cause the WQI to flip bewteen Idle and In Progress status [Defect #10259] - Receive Work Items complete successfully when images are refused due to disk space issues
Updates for ticket #10259 caused this issue.  We just had to remove a check
when the TotalFilesToProcess was set to 1 that it was returning that the study
was stil processing.  This was causing us to always Idle and set the ProcessTime
to the Expiration Time, so it would continually reschedule itself and never
complete.  The refactoring of the patch for #10259 makes this no longer required
since we Idle at least once when we have the no disk space available condition.