[Defect #10272] - Excessive subscriptions to WorkItemServiceActivityMonitor will degrade performance and leak memory over time
Introduced a new Cleanup method on IDicomScp.  This is called once when an
association is closed/released/aborted.  The Cleanup() method in turn for the
SCP extensions now removes the events on the ActivityMonitor and then disposes
the monitor itself.  Hopefully this should help with the release of the memory
association with the activity monitors.