[Requirement #10273] - ImageServer shall function with .NET 4.
Updates to make the IS work with .NET 4.0.  Managed to update the AJAX toolkit
to use the 4.0 version included with .NET.  We no longer are dependent on
that external project!  Also, had to make soem changes to support the new
licensing being stored through the enterprise server.  Finally, updates
needed to be done to all the javascript files that created classes to match
server side classes, some code to check if the javascript was already created
was no longer needed after the ajax upgrade, in fact it caused the code
to not work.  As part of this, cleaned up the js files also.  Finally, upgraded
to the latest version of the AjaxControlToolkit.