[Defect #9594] - Web GUI Study queyr very slow when large number of partitions are created
Introduced new ServerPartitionSelector control that changes the page so it goes to
the server whenever you switch between partitions.  This makes only a single copy of the gridview
on the page, an thus should be expandable to a large number of partitions.  note that this has only been applied to the
study page, and not the others.

Introduced 3 new log statements to diagnosis the problem:  added a log as to
where each file is stored when it comes in, added a log when a study is being
verified as to where it is being worked on, and added a log after a WorkItem is
completed as to where it it stored.