[Requirement #10258] - The workstation shall include a Simple Order Entry screen, launched from calendar view
Did some refactoring in order to move the "Additional Info" page out of the core plugin and into the Workflow.Extended plugin.  In order to do this, the page had to be converted into an extension page in every context in which it was used, which was quite a bit of work.