[Requirement #10273] - ImageServer shall function with .NET 4.
Discovered a bunch of issues with the Webstation not working, and cleaned up.
Most issues were related to the Marmot refactoring, and updates required to
make that work.  Had to implement several service providers related to local
configuration and the server tree.  Also some changes were required to get
the streaming study loader working properly, including updates to the
web.config.

Note that a new update may be required to the build process to pdate a web.config
configuration option related to the port that the streaming service is listening
on for the IS.