[Defect #10144] - Retrieve Requests never fail when SOP rejected & can't be sent
The callback that was determining when to send the final response message
was being called before the failure/remaining sub operation variables were being
incremented/decremented.  This was causing the code to never send the final
response.  Also, added some defensive code when the association is shutting
down to send a final response, if one hasn't been sent yet.  This shouldn't be
needed, but there's a chance some sort of exception could occur that would cause
this cleanup to not work properly.