[Requirement #10412] - When an addendum is created, the encapsulated PDF stored in the study must be updated to include the addendum
Service updates for handling duplicate reports. The basic code has been implemented
but not unit tested yet.  We now will accept duplicate reports (Encapsulated PDF
or Encapsulated CDA) if the Instance Creation Date/Time tags in the header are
for a newer time than the existing file.  The existing file is simply
overwritten in this case, otherwise the duplicate is deleted.  Note that the
new partition setting overrides the duplicate SOP policy for the partition
when it is enabled, otherwise we drop back to the duplicate sop policy.