[Requirement #10412] - When an addendum is created, the encapsulated PDF stored in the study must be updated to include the addendum
Updates to duplicate handling.  Added code to handle updating the study xml
file when a duplicate report is received.  Also made the code so it ignores
reports if the Date/Time is the same as a prior report.  Finally, made updates
so the folders where the duplicate is stored are cleaned up.