[Requirement #10229] - The workstation will provide a new Scheduling folder system, with a calendar view for scheduling patient appointments
Fixed an issue that would only manifest if the Scheduling FS was configured to be the first one.  In this case, when you first opened the homepage, the Scheduling FS would be selected but instead of showing the calendar view it would show the folder contents view.  Clicking on any scheduling folder would then cause exceptions.  This was all due to a bug in StackedComponentContainer, fixed here.

 * also corrected some of the string resources in the Calendar.