[Requirement #10255] - The system will provide a PDF-based Print Report feature to replace the existing (and broken) IE-based Print feature
Discovered an issue during integration testing.  Because each shred is hosted in its own app domain, the RIS server will actually instantiate two instances of the Http Proxy server: one for the application server (for manually printing reports) and one for the publication shred (for sending pdf's to the imageserver).  For this reason, we actually need to specify at least two free ports that the proxy server can listen on.  Changed the settings to specify a port range.