[Requirement #10411] - Radiology reports will be published to the ImageServer to be included in the associated study as an encapsulated PDF
Refactored the publication shred so that instead of processing publication actions directly, it uses a two-step process whereby processing of the PublicationStep consists merely of queueing up work items for each publication action extension, and the second step is to go and process those work items.  This is far more robust, since it isolates failure of a given publication action so that it does not affect the others.

Incidentally, this will also fix the bug where reports were being published to the ImageServer with a status that said incorrectly "This report has not yet been verified".  Because of the two-step processing now used, things will now be done in correct order.