[Requirement #10142] - The application should support manually attaching documents (PDF) to patients and orders
The AttachedDocumentPreviewComponent no longer provides an embedded preview, since this feature a) will not work at all on x64 util Adobe releases a compatible ActiveX control, and b) did not work on x86 even though it is supposed to.

Instead, it just lists the attachments, and user must double-click to view in an external application.  Also added an explicit "view" action on the toolbar, in case the double-click isn't obvious.