[Defect #9650] - DICOM SDK: Typos in certain tags containing the character "mu" (SI prefix micro) [Defect #9824] - Ability to add additional SOP Classes to ImageServer through DB not supported [Requirement #10421] - Upgrade the DICOM Toolkit Data Dictionary to the 2011 version
Updated to the 2011 version of the standard.  Updated the data dictionary generator
such that the SopClass generation takes into account the issue with #9824.  Also
updated the generator to translate the mu character to U.

Also made some general updates with the dictionary generation code to clean it
up and address some resharper issues.  Finally, checked in the 2011 version of the
standard in xml.