[Defect #10456] - Wildcard query does not return results when no value for tag
Upon review, there were issues in both the GUI and the DICOM service processing
queries.  Adjusted the QueryHelper.SetStringCondition() routine to not add
a search condition when there were only wildcards for the tag.  Also, added a
new QueryHelper.SetGuiStringCondition() method to be used by the GUI.  Changed the
AlertDataSource, ApplicationLogDataSource, DeletedStudyDataSource, and the StudyDataSource
to use the new method.  Also made updates to the Devices, Filesystems, Server
Partition and Partition Archive configuration screens to address the issue
and use the new helper routines when searching for text attributes.  Note that
most of these pages were not allowing null values for tags, but the code is still
a bit cleaner in each case.