[Requirement #10312] - User shall be warned about collisions
Corrected an issue with the implementation of the in-memory "Between" operator.  It was implemented as

start <= x < end

but in order to match the database (MS-SQL Server) implementation, it should be implemented as

start <= x <= end

Note: ultimately didn't end up using the Between operator to implement this TR, but decided to commit the fix in any case.  There is minimal risk here because the in-memory operators are thus far only used in the OWLS code and the above change in implementation will not have any practical impact on the OWLS functionality.