[Requirement #10312] - User shall be warned about collisions
DB Schema change: add ScheduledEndTime_ (datetime) to Procedure_

In order to perform the database queries necessary for collision detection, the Procedure needs to store a scheduled end time, in addition to the start time and duration (because even though the end time is just the start time + duration, this cannot be efficiently queried).
 * added ScheduledEndTime to Procedure
 * added index to dbi file

Also, in order to make collision detection easier at the boundary cases, we need to be scheduling procedures with minute-level resolution - that is, the seconds and milliseconds components of the times need to be discarded before the object is saved in the database.
 * modified Procedure.Schedule(...) so that it truncates the time to minute-level resolution
 * updated corresponding unit tests