[Defect #8940] - JPEG Lossless Process 14 codecs corrupt pixel data when Bits Stored=8, Bits Allocated=16
Updated Jpeg code to convert 16 bits allocated, 8 bits stored pixel data to
8 bit data, before conmpression.  The previous code was just compressing, and
was compressing the first 1/2 of each image row.  We update the
Bits Allocated to 8 after the conversion.

As part of this, a new method was added to DicomPixelData to change a 16 bit
image to 8 bits, which only works when BitsAllocation=16/BitsStored=8.