[Defect #9691] - Database Deadlock Errors whne deleting with large databases
Added the recommended index into the database based on the forum support.  Note
that in testing we were unable to replicate the issue.  The issue would require
a large number of edits of studies comparied to the number of studies stored on
a system, since the problem was related to the StudyHistory table.