[Defect #10318] - Unable to delete empty partition when study count is negative
Made an attempt to prevent the negative study count.  I think this was caused
by an exception when doing the delete that caused the delete stored procedure
to be called more than once.  It might be linked to deleting studies for the
same patient at the same time.  This probably doesn't have to be tested besides
verification of basic study count functionality.

Also, changed the check to see when a partition can be deleted to allow
negative values for study count.