[Requirement #10449] - The system shall allow an administrative user to log out another user
Made a bunch of improvements to allow the application to handle the situation where it's current session has been killed remotely by an admin.
 * Added an ISessionManager.InvalidateSession() method, which essentially marks the current session as invalid, requiring the user to have to re-authenticate in order to continue using the application.
 * Added a Desktop.Application.InvalidateSession() method, which exposes this functionality publicly to the application.
 * Adjusted one of the RIS exception policies so that it calls InvalidateSession().
 * Ris.Client.SessionManager does not yet implement InvalidateSession(), since it isn't required for the current project.

The net effect is that, if a RIS user tries to execute a command when their session has been terminated by an admin, they are simply shown a login screen and can re-enter their password and continue working.