[Requirement #10416] - The RIS will be integrated into the Team Edition workstation
The RIS client must behave reasonably well in offline mode.  In order to achieve this, decided to separate out the handling of EndpointNotFoundException from the CommunicationExceptionPolicy into its own policy (EndpointNotFoundExceptionPolicy) so that we can present the user with a message box that explicitly mentions that the server is known to be offline.