[Requirement #10411] - Radiology Reports will be published to the ImageServer
Fixed an issue where the following exception would be thrown when shutting down the shred host:
"Unexpected exception stopping shred (shred still running): Publication
System.CannotUnloadAppDomainException: error while unloading appdomain"

It was caused because the HTTP Print server was not being shutdown, hence a thread was still active when the app domain was unloaded.  Refactored code so that the HTTP Print server is started and stopped with each print job, rather than running continuously, to minimize chances of this error happening.