[Requirement #10416] - The RIS will be integrated into the Team Edition workstation
Looked for instances where LoginSession.Current was assumed to be non-null, and added in null checks in select places since this assumption is not valid in Team Edition.
(Didn't add null checks everywhere, because there are many places where you can't even get that far if there is no current LoginSession).