[Requirement #6418] - Reassigning a report that's in transcription causes the report to fall off the transcription worklist
The defect is due to a design flaw with the Reassign command in the Radiology Admin folder. The command only allows you to re-assign the item to another radiologist, but it doens't make sense to re-assign a transcription step to a radiologist.

Best fix I could come up with (barring re-designing the whole thing) was to disable the Reassign command for transcription items.

Also noticed that the Discard Report command was disabled when it should have been enabled.  Made a code change to fix this, without being terribly sure of the implications.  Bottom line is this entire area of the code is junk, and we just have to hope the duct tape and bubble gum hold out until we can reach the getaway car.