[Requirement #8151] - Ensure the source code repository is managed correctly
* Fixed issue in the WebPortal which surfaced when Marmot was merged into Yen. In Yen, users could not open a study if it contains a report because an NullReferenceException would be thrown by the PdfHttpHandler.
The issue was caused related to the ImageServerServerDirectory class (which replaced other classes when we moved to .NET 4). This class checks for thread principal to perform fine-grain access.
When report was accessed via PdfHttpHandler, the thread principal was not set up properly which led to the NullReferenceException later in the code on this line:

var uri = new Uri(string.Format("http://{0}:{1}/wado", ae.ScpParameters.HostName, ae.StreamingParameters.WadoServicePort));