[Ticket #10550] - Null reference when performing multiple procedures
Decided to fix this at the root cause level rather than just fixing the symptoms described in the ticket.  Modified the implementation of ItemCollection.Replace(...) so that it computes all rows that need updating prior to updating any rows.  This should prevent weird interactions of the kind that caused this defect.