[Ticket #10527] - MWL should return a study instance UID that is stored in the RIS system
Fixed an issue where adding a new procedure to an existing order would throw an exception because the StudyInstanceUID field wasn't being set.  Refactored code so that the study instance UID must be passed into the Procedure constructor, so this kind of error can't be made in the future.

Also, for the sake of consistency with the way accession numbers and procedure numbers are generated, refactored the UID generation into a broker.  This may seem like an odd choice, since UID generation doesn't come from the RIS database and arguably has nothing to do with brokers, but it does give consistency to the way unique identifiers are generated and assigned to entities in the RIS, and allows for unit testing with a mock UID broker.