[Requirement #10540] - Basic imaging tools will be available in the toolbar
Got deserialization of the Message object working.  Its a bit of a hack, but
it works.  Need to investigate if there's alternates for this further.

Also, have some basic functionality with clicking on buttons working, but note
that it ends up failing after awhile and causing exceptions on the server,
need to investigate this a bit further.

Also, introduced a simple method for writing to the console/logging so we
can shut this off/turn it on easily if we want.  Could be extended further in
the future.