[Requirement #10564] - In RIS preview pages, images do not appear on the page
Turns out this is caused by a block of CSS that was being applied globally to *all* <img> tags, preventing ALL IMAGES from showing up.  The CSS was added as part of ticket #1546, and apparently had something to do with preventing IE from rendering transparent images with a "blue tinge".  However, given that it now prevents images from showing up at all, I think a blue tinge is the lesser evil.  Commented out the code for now, in case it turns out that we still need it.