[Requirement #10258] - The workstation shall include a Simple Order Entry screen, launched from calendar view
Because the Simplified Order Entry screen allows selecting a patient, there is no longer any reason to disable the New Order tool when an item is not selected.  The New Order tool is now always enabled, in both the Registration FS and the Patient Search screen.  If an item is selected, the patient field will be pre-populated.  Otherwise, the patient must be selected on the Order Editor screen.