[Defect #10544] - Most of the ImageServer admin pages are broken in IE8
Problem was caused by a bug in the AJAX control tookit, specifically in the
June 2012 release.  Downloaded the source for this release, and applied the
patch on this page:

http://ajaxcontroltoolkit.codeplex.com/workitem/27053

Note that I also double checked the Hungarian bug from ticket #9750 reported here:

http://ajaxcontroltoolkit.codeplex.com/workitem/23221

and I think this was already resolved in the June 2012 version.