[Requirement #10258] - The workstation shall include a Simple Order Entry screen, launched from calendar view
Improved the title bar of the dialog when editing or replacing an order, so that it actually includes the A# of the order being edited.

Note: the exception to this is when the editor is launched from the calendar view, in which case it just says "Modify Procedure(s)", because the information required for the title bar isn't available.  This would be a bit more work to fix, and we don't have time.