[Ticket #10602] - User should be warned when scheduling a procedure in the past
We now check for procedures being scheduled in the past and warn the user.  Note that only *modifiable* procedures are considered in this check.  Procedures that have been cancelled, or are pending cancellation, or that have already been started, are not included in the check, since their scheduled times cannot be modified in any case.