[Requirement #10593] - User messages, including alerts, will be shown in message boxes [Requirement #10381] - Ensure the source code repository is managed correctly
Minor code cleanup, updated comments and copyrights in some of the javascript
files.

Also, changed the message box display so that localized strings for the
dialog labels are passed from the server. Finally, added a mechanism to
include translated strings that are hidden in the Default.aspx page, that
can in turn be grabbed by the javascript and inserted into dialogs shown to
the user.  These translated strings are stored in a Webstation.rex file in the
web application.