[Requirement #10590] - The user can stack through images using a scroll bar
This is finally the final change. Previously, I had been trying to save
scroll values to prevent circular updates, but there were still some problems
when there was only 2 frames in an imagebox.  Instead, made a server side update
so that when a propertychanged event for the TopLeftPresentationImageIndex
property is sent to the server, we don't in turn send a property changed event
for that same change back to the client.  Also, added a simple check on the
client when the TopLeftPresentationImageIndex changed event is sent from the
server, we don't update back to the server after the change.  This resolves
the flakyness.