[Requirement #10699] - The user can quickly capture key images
Reworked the code into a new KeyImageDisplaySet class and removed it from
KeyImageClipboard for creating the key images. Changed the KeyImageDisplaySetDescriptor
to inherit from DisplaySetDescriptor instead of DicomDisplaySetDescriptor so that
it could be study based instead of series based.  Reworked so that if multiple
key images are created, they're placed in the same display set per study.